Power optimiation.
Use hardware LCD blanking to show the blank page in the countdown. This is very power efficient because (1) we only need to change a single bit to show the blank screen rather than writing all of the segments to be blank, and (2) we can now *always* keep the HHMMSS page rendered in the LCDMEM bank so we only ever need to write the digits that actually change. I think this update code is now pretty damn close to being algorithmically optimal! The only way to make it more efficient now is to avoid the crappy compiler and implement in ASM.